move SQL related functions from Herder to HerderPersistence class
